+++ config/config.exs
@@ -1,9 +1,3 @@
 use Mix.Config
 
-config :airbrake_client,
-  api_key: {:system, "AIRBRAKE_API_KEY", "FAKEKEY"},
-  project_id: {:system, "AIRBRAKE_PROJECT_ID", 0},
-  host: {:system, "AIRBRAKE_HOST", "https://airbrake.io"},
-  private: [http_adapter: HTTPoison]
-
 import_config "#{Mix.env()}.exs"